requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pymongo==4.16.0
python-dotenv==1.2.2
google-genai==1.66.0
streamlit==1.55.0
requests==2.32.5
pandas==2.3.3
numpy==2.4.3
pydantic==2.12.5